Dev Update: Added functionality to enable/disable comments feature

+++ pr-deploy.sh
@@ -8,6 +8,11 @@ PID_FILE="/srv/pr-deploy/nohup.json"
 COMMENT_ID_FILE="/srv/pr-deploy/comments.json"
 
 comment() {
+    # Check if comments are enabled
+    if [ "$COMMENT" != true ]; then
+        return
+    fi
+    
     local status_message=$1
     echo $status_message
 
@@ -72,14 +77,16 @@ if [ ! -f "$PID_FILE" ]; then
     echo {} > $PID_FILE
 fi
 
-# Initialize the JSON file for comment if it doesn't exist
-if [ ! -f "$COMMENT_ID_FILE" ]; then
+# Initialize the JSON file for comment if comments are enabled and it doesn't exist
+if [ "$COMMENT" == true ] && [ ! -f "$COMMENT_ID_FILE" ]; then
     echo {} > $COMMENT_ID_FILE
 fi
 
-# Handle COMMENT_ID
-COMMENT_ID=$(jq -r --arg key $PR_ID '.[$key] // ""' ${COMMENT_ID_FILE})
-comment "Deploying ⏳"
+# Handle COMMENT_ID if only comments are enabled
+if [ "$COMMENT" == true ]; then
+    COMMENT_ID=$(jq -r --arg key $PR_ID '.[$key] // ""' ${COMMENT_ID_FILE})
+    comment "Deploying ⏳"
+fi
 
 # Ensure docker is installed
 if [ ! command -v docker &> /dev/null ]; then
